% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo_download.R
\name{etopo_download}
\alias{etopo_download}
\title{Download ETOPO 2022 Bathymetery data}
\usage{
etopo_download(resolution = 60, force = FALSE)
}
\arguments{
\item{resolution}{numeric resolution in arcsecs (one of 30, or 60). Defaults to 60 arcsecs.}

\item{force}{Logical. If data has previously been downloaded, it will force a new
download and update of the OSM data.}
}
\description{
Downloads bathymetery data for map making.
}
\references{
https://osmdata.openstreetmap.de/data/land-polygons.html
}
\author{
Devin S. Johnson
}
