% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo_rast.R
\name{etopo_rast}
\alias{etopo_rast}
\title{Extract Bathymetry Data From Downloaded ETOPO 2022 data.}
\usage{
etopo_rast(x, resolution = 60, ...)
}
\arguments{
\item{x}{an \code{sf} spatial object that will define the extent of the bathymetry raster.}

\item{resolution}{Resolution of the ETOPO bathymetry raster, either \code{60} or \code{30} arc-seconds. Defaults to \code{resolution = 60}.}

\item{...}{Addtional arguments passed to \code{terra::\link[terra]{project}}.}
}
\description{
The majority of this function is taken from the \code{topotools} package here: \url{https://github.com/BigelowLab/topotools/tree/main}. It was modified to work within
the `picMaps` package.
}
\references{
\itemize{
\item NOAA National Centers for Environmental Information. 2022:
ETOPO 2022 15 Arc-Second Global Relief Model. NOAA National Centers for
Environmental Information. DOI: 10.25921/fd45-gt74. Accessed 2023-03-17.
\item \url{https://github.com/BigelowLab/topotools/tree/main}
}
}
