% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_storage.R
\name{get_data_loc}
\alias{get_data_loc}
\title{Show picMaps Data Storage Directory}
\usage{
get_data_loc()
}
\description{
This function provides the location where `picMaps` is storing and retrieving spatial data.
}
