% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_funs.R
\name{get_grad}
\alias{get_grad}
\title{Find gradient field of a elevation raster}
\usage{
get_grad(r, scale = TRUE)
}
\arguments{
\item{r}{SpatRaster object}

\item{scale}{Logical, Should the gradient be scaled by resolution of the coordinates. Defaults
to `scale = TRUE`}
}
\description{
Find gradient field of a elevation raster
}
