% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_coast.R
\name{hawaii_coast}
\alias{hawaii_coast}
\title{Coast polygons for Main Hawaiian Islands}
\usage{
hawaii_coast(keep = 0.5, union = TRUE, crs = 4326)
}
\arguments{
\item{keep}{see \code{\link[picMaps]{osm_coast}}}

\item{union}{see \code{\link[picMaps]{osm_coast}}}

\item{crs}{A coordinate system object, See \code{\link[sf]{st_transform}}. Defaults to
ESPG = 4326 with 0-360 longitude values.}
}
\description{
Coast polygons for Main Hawaiian Islands
}
