% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_funcs.R
\name{hi_benthic_habitat}
\alias{hi_benthic_habitat}
\title{Import benthic habitat data for the Main Hawaiian Islands}
\usage{
hi_benthic_habitat()
}
\description{
Import benthic habitat data for the Main Hawaiian Islands
}
