% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_funcs.R
\name{hi_watersheds}
\alias{hi_watersheds}
\title{Watersheds}
\usage{
hi_watersheds(type = "DAR")
}
\arguments{
\item{type}{DAR version (`type="DAR"`; default) or `type="regular"`.}
}
\description{
Watersheds
}
