% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_coast.R
\name{osm_coast}
\alias{osm_coast}
\title{Get A Coastline \code{sf} Polygon Object For Plotting and Mapping}
\usage{
osm_coast(x, keep = 0.2, union = FALSE)
}
\arguments{
\item{x}{An \code{sf} spatial object. The coastline will be cropped to the bounding box.}

\item{keep}{The amount of data retained after simplification with \code{\link[rmapshaper]{ms_simplify}}}

\item{union}{Logical. Should the returned object be returned as a single \code{sf} \code{MULTIPOLYGON} object}
}
\description{
Uses downloaded OSM data for constructing an \code{sf} polygon coastline data object.
Prior to using this function you must run \code{\link[picMaps]{download_osm}}.
}
\author{
Josh M. London and Devin S. Johnson
}
