% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_download.R
\name{osm_download}
\alias{osm_download}
\title{Download Open Street Map Coastline Shapefile from Daylight Map Distribution}
\usage{
osm_download(force = FALSE)
}
\arguments{
\item{force}{Logical. If data has previously been downloaded, it will force a new
download and update of the OSM data.}
}
\description{
Downloads OSM data for plotting telemetry data and use distributions.
}
\references{
https://osmdata.openstreetmap.de/data/land-polygons.html
}
\author{
Josh M. London and Devin S. Johnson
}
