% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_install.R
\name{osm_install}
\alias{osm_install}
\title{Download Open Street Map Coastline Shapefile from Daylight Map Distribution}
\usage{
osm_install(zip_file, force = FALSE, clean_shp = TRUE)
}
\arguments{
\item{zip_file}{path to the OSM data .zip file `land-polygons-complete-4326.zip`. If left unspecified, the user will see a directory browser to navigate to it.
See the url in the references to dowload it by hand. Also, see `\link{osm_download}` to download and install it.}

\item{force}{Logical. If data has previously been downloaded or installed, it will force a new
install and update of the OSM data.}

\item{clean_shp}{Logical. If `TRUE` the original shape file will be deleted and only the `.gpkg` file will be retained.}
}
\description{
Downloads OSM data for plotting telemetry data and use distributions.
}
\references{
https://osmdata.openstreetmap.de/data/land-polygons.html
}
\author{
Devin S. Johnson and Josh M. London
}
