% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_funs.R
\name{raster_fill}
\alias{raster_fill}
\title{Fill raster holes and NAs with local interpolation}
\usage{
raster_fill(r)
}
\arguments{
\item{r}{SpatRaster object}
}
\description{
Fill raster holes and NAs with local interpolation
}
\details{
Uses focal means to infill holes and NAs
}
