% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_storage.R
\name{set_data_storage}
\alias{set_data_storage}
\title{Set Data Storage Directory}
\usage{
set_data_storage(path = NULL)
}
\arguments{
\item{path}{A directory path specifying where spatial data will be stored for the package}
}
\description{
The directory specified by \code{path} will be used to store downloaded spatial data. By storing the
data in a separate directory from the installed package it will not need to be re-downloaded when the package
is updated. You simply need to rerun this function to let the package functions know where the data is located.
}
