% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_addons.R
\name{st_divide_poly}
\alias{st_divide_poly}
\title{Divide a polygon into k equal-area polygons}
\usage{
st_divide_poly(polygon, k, n_samples = 10000)
}
\arguments{
\item{polygon}{The `sf` polygon object to be divided}

\item{k}{The number of resulting polygons}

\item{n_samples}{The number of sample points to perform k-means clustering and divide the polygon.}
}
\description{
Uses k-means clustering and Voronoi tessellation to divide a polygon
into k equal-area polygons.
}
\author{
Google Gemini
}
