% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_addons.R
\name{st_overlap_dateline}
\alias{st_overlap_dateline}
\title{Check if spatial data bounds overlap dateline}
\usage{
st_overlap_dateline(x)
}
\arguments{
\item{x}{An `sf` spatial data object}
}
\description{
Check if spatial data bounds overlap dateline
}
