% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_addons.R
\name{st_partition_by_inner_polygons}
\alias{st_partition_by_inner_polygons}
\title{Divide a polygon into Voronoi tessellations based on closeness to inner polygons instead of points}
\usage{
st_partition_by_inner_polygons(
  outer_poly,
  inner_polys,
  inner_id,
  sample_density = 5000
)
}
\arguments{
\item{outer_poly}{The `sf` polygon object to be divided}

\item{inner_polys}{An `sf` data frame containing inner polygons to partition the outer polygon}

\item{inner_id}{A column in `inner_polys` which will label the outer polygon partitions.}

\item{sample_density}{The density of points used to fill the inner polygons
to perform the partitioning. Defaults to `sample_density = 5000`.}
}
\description{
Divide a polygon into Voronoi tessellations based on closeness to inner polygons instead of points
}
\author{
Google Gemini and Devin S. Johnson
}
