% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_funcs.R
\name{fetch_cms_phys_3d}
\alias{fetch_cms_phys_3d}
\title{Download Copernicus Marine Service physical oceanography data (3d variables)}
\usage{
fetch_cms_phys_3d(
  region,
  variable,
  timerange,
  verticalrange = c(0, -0.5),
  return_type = "stars",
  progress = FALSE,
  asset = "ARCO",
  ...,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{region}{xxx}

\item{variable}{xxx}

\item{timerange}{xxx}

\item{verticalrange}{xxx}

\item{return_type}{One of `"stars"` or `"raster"`. A `stars` object is the default.}

\item{progress}{xxx}

\item{asset}{xxx}

\item{...}{xxx}

\item{username}{xxx}

\item{password}{xxx}
}
\description{
Download Copernicus Marine Service physical oceanography data (3d variables)
}
\author{
Devin S. Johnson
}
