% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_funcs.R
\name{fetch_cms_phys_3d}
\alias{fetch_cms_phys_3d}
\title{Download Copernicus Marine Service physical oceanography data (3d variables)}
\usage{
fetch_cms_phys_3d(
  region,
  variable,
  layer,
  timerange,
  verticalrange = 0,
  return_type = "raster",
  progress = FALSE,
  asset = "ARCO",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{region}{An `sf` or `bbox` object. The bounding box will be used as the bounds for CMS
data subsetting and download}

\item{variable}{Variables from the GLOBAL_MULTIYEAR_PHY_001_030 product for download. Must be
any subset of `c("thetao","so","uo","vo")` any other variables will be ignored.}

\item{layer}{One of `"daily"` or `"monthly"` for either temporal resolution.}

\item{timerange}{A 2 vector of start and end dates. Both must be POSIX.}

\item{verticalrange}{Range of depth values to download, e.g., `c(0,-100)` will download
all CMS depth layers between 0-100m. Default is `0` for surface layer.}

\item{return_type}{One of `"stars"` or `"raster"`. A `stars` object is the default.}

\item{progress}{Output process of download.}

\item{asset}{See `?CopernicusMarine::cms_download_subset`}

\item{username}{Copernicus Marine Service account username,
see `?CopernicusMarine::cms_download_subset` and `?CopernicusMarine::cms_get_username`}

\item{password}{Copernicus Marine Service password,
see `?CopernicusMarine::cms_download_subset` and `?CopernicusMarine::cms_get_password`}

\item{...}{Ignored arguments}
}
\description{
Download Copernicus Marine Service physical oceanography data (3d variables)
}
\author{
Devin S. Johnson
}
