% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_funcs.R
\name{fetch_cms_phys_surface}
\alias{fetch_cms_phys_surface}
\title{Download Copernicus Marine Service physical oceanography data (surface values)}
\usage{
fetch_cms_phys_surface(
  region,
  variable,
  layer,
  timerange,
  return_type = "raster",
  progress = FALSE,
  asset = "ARCO",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{region}{An `sf` or `bbox` object. The bounding box will be used as the bounds for CMS
data subsetting and download}

\item{variable}{Variables from the GLOBAL_MULTIYEAR_PHY_001_030 product for download.}

\item{layer}{One of `"daily"` or `"monthly"` for either temporal resolution.}

\item{timerange}{A 2 vector of start and end dates. Both must be POSIX.}

\item{return_type}{One of `"stars"` or `"raster"`. A `stars` object is the default.}

\item{progress}{Output process of download.}

\item{asset}{See `?CopernicusMarine::cms_download_subset`}

\item{username}{Copernicus Marine Service account username,
see `?CopernicusMarine::cms_download_subset` and `?CopernicusMarine::cms_get_username`}

\item{password}{Copernicus Marine Service password,
see `?CopernicusMarine::cms_download_subset` and `?CopernicusMarine::cms_get_password`}
}
\description{
Download Copernicus Marine Service physical oceanography data (surface values)
}
\author{
Devin S. Johnson
}
