% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_coast.R
\name{hawaii_coast}
\alias{hawaii_coast}
\title{Coast polygons for Main Hawaiian Islands}
\usage{
hawaii_coast(keep = 1, union = TRUE)
}
\arguments{
\item{keep}{see \code{\link[picMaps]{osm_coast}}}

\item{union}{see \code{\link[picMaps]{osm_coast}}}
}
\description{
Coast polygons for Main Hawaiian Islands
}
