% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predefined_coast.R
\name{mariana_coast}
\alias{mariana_coast}
\title{Coast polygons for the Mariana Islands}
\usage{
mariana_coast(keep = 0.5, union = TRUE, crs = 6637)
}
\arguments{
\item{keep}{see \code{\link[picMaps]{osm_coast}}}

\item{union}{see \code{\link[picMaps]{osm_coast}}}

\item{crs}{A coordinate system object, See \code{\link[sf]{st_transform}}. Defaults to
ESPG = 6637, the Guam map grid projection (UTM zone 55N).}
}
\description{
Coast polygons for the Mariana Islands
}
